requirements.txt
# Core dependencies
openai==1.3.0
python-dotenv==1.0.0
streamlit==1.31.0
pydantic==2.5.3

# Optional performance improvement
watchdog

# PROJECT TWO
ollama
pydantic-ai

